refactor: move DbUpdateExceptionExtensions to BuildingBlocks and use extension method for registration validation
